fix(GoalTableList): move open preview click handler inside `<GoalTableList />`
